libaur: Split up updatepkgs more
This will let people extend it to use different executable names in case
their distribution does not call pacman 'pacman', or they do not
actually have /usr/bin/paclist.

I could have a variable that use for the binary locations, but this
allows you to pass the same output and still get the same results.