pywer: Add formatted printing to the mix
Right now we don't support printf formatting, but you can still do most
things that people use. Also, the '\n' at the end of fmt is not implied,
if you want it to be on multiple lines, you need to put a '\n' at the
end of the string.

Also updated the documentation a bit.

One of the function names changed as well, pretty_print_simple_info is
now pretty_print_info, since it does a whole lot more than just the
simple_info part.