printer: Change how we print multiple returns
Previously we just +='d the string we were going to print, which forced
me to do two separate things for both pretty_print_search and
pretty_print_info, and with both of them we want to put a '\n' between
the results, but not have an extra one at the end. This was working for
info, but I hadn't bothered for search. Now there is a new way of
getting the final printed string. Instead of adding to a string, we
append the string to a list of strings that represent the results and
then use str.join() to do that without any fancy logic in our library.