printer: More fixing of the printing
Right now, '\n' in --format doesn't actually put a newline, but an
actual '\n'. You can still use this with `echo -e $()` or something. You
can actually pass an actual newline using a multiline string in your
shell and it will print a newline, but the \n is treated like a string
at the moment.

It uses a specified sep variable depending on what kind of printing will
be taking place (formatted or unformatted) and then uses that for the
str.join().

Also, the optdepends array has to be processed separately because we
don't want to use wrapper.fill() on it.