Make this work with python 2
Remove the explicit 'python3' declaration in the #!, lists don't have
the .copy() function in them, and we have to change the wway we pick
things out of the config file. Most of it worked out of the box with
python 2.7.6, but now it all does, so everything should be working if
you only have access to python2 or want to install this with python2.