More pep8 compliance
This isn't a real code change, more just a change in how the code is
written. If it doesn't pass a pep8 test then it's not good. Going to set
up a pep8 and `./setup.py test && ./setup.py check` thing for each pull
request and as a pre-commit hook.